Clean up Annotator and Store plugin tests
Major structural changes to the core Annotator object and its
relationship with the store plugin have broken a lot of (fragile and
delicately coupled) tests. I've been quite aggressive in removing tests
that cross what I consider to be unit boundaries, or which probe the
implementation rather than the interface of the unit they are supposed
to be testing. In particular, the Store plugin test suite looks much
more sensible now.